ml-services/requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
prophet
torch
torchvision
pillow
